refactor: 홈 구성 재배치
