+++ app/(anon)/page.tsx
@@ -1,32 +1,11 @@
-'use client';
+import { Suspense } from 'react';
+import Home from './home/page';
 
-import { useSearchParams } from 'next/navigation';
-import { Suspense, useEffect } from 'react';
-import { useLoginModal } from '@/app/hooks/useLoginModal';
-import { KakaoLoginModalContainer } from '@/app/components';
-
-export default function Home() {
-  const searchParams = useSearchParams();
-  const { showLoginModal, handleOpenModal, isAuthenticated, loading, handleCloseModal } =
-    useLoginModal();
-
-  const handleLogin = () => {
-    const kakaoAuthUrl = `https://kauth.kakao.com/oauth/authorize?client_id=${process.env.NEXT_PUBLIC_KAKAO_REST_API_KEY}&redirect_uri=${process.env.NEXT_PUBLIC_KAKAO_REDIRECT_URI}&response_type=code`;
-    window.location.href = kakaoAuthUrl;
-  };
-
-  useEffect(() => {
-    if (searchParams.get('login') === '1' && !isAuthenticated && !loading) {
-      handleOpenModal();
-    }
-  }, [searchParams, isAuthenticated, loading, handleOpenModal]);
-
+export default function Page() {
   return (
     <>
       <Suspense fallback={<div>로딩 중...</div>}>
-        {showLoginModal && (
-          <KakaoLoginModalContainer onLogin={handleLogin} onClose={handleCloseModal} />
-        )}
+        <Home />
       </Suspense>
     </>
   );
